requirements.txt
pyttsx3==2.90
PyAudio==0.2.11
SpeechRecognition==3.8.1
wikipedia==1.4.0
Pillow==8.40